♻️ refactor: api, service 폴더 분리
